requirements.txt
PySide6
pandas
openpyxl
rapidfuzz
loguru
Pillow
pyinstaller